requirements.txt
aenum==3.1.15
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
attrs==24.1.0
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Deprecated==1.2.14
Flask==3.0.3
frozenlist==1.4.1
future==1.0.0
gunicorn==22.0.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
line-bot-sdk==3.11.0
MarkupSafe==2.1.5
multidict==6.0.5
packaging==24.1
paho-mqtt==2.1.0
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
requests==2.31.0
six==1.16.0
typing_extensions==4.12.2
urllib3==2.2.2
websockets==12.0
Werkzeug==3.0.3
wrapt==1.16.0
yarl==1.9.4
